+++ src/__mocks__/data/economics.ts
@@ -1,0 +1,11 @@
+export const economics = {
+  totalSupply: 24352795,
+  circulatingSupply: 24352795,
+  staked: 2319671,
+  price: 8.4,
+  marketCap: 204563478,
+  apr: 0.664383,
+  topUpApr: 0.11788,
+  baseApr: 1.518661,
+  tokenMarketCap: 46044832
+};